+++ app/build.gradle
@@ -19,9 +19,14 @@ android {
     }
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
+    }
+    flavorDimensions "datasource"
+    productFlavors {
+        firebase { dimension "datasource" }
+        twitter { dimension "datasource" }
     }
 }
