+++ app/src/main/java/edu/gwu/androidtweets/IntentKeys.kt
@@ -1,0 +1,5 @@
+package edu.gwu.androidtweets
+
+object IntentKeys {
+    val KEY_ADDRESS = "address"
+}
