+++ app/src/main/java/edu/gwu/androidtweets/SharedPrefsKeys.kt
@@ -1,0 +1,6 @@
+package edu.gwu.androidtweets
+
+object SharedPrefsKeys {
+    val KEY_USERNAME = "username"
+    val KEY_PASSWORD = "password"
+}
